requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
python-dotenv==1.0.1
PyGithub==2.6.1
setuptools==80.9.0
spacy==3.7.5
textstat==0.7.5
vaderSentiment==3.3.2
sentence-transformers==3.4.1
bertopic==0.16.4
scikit-learn==1.5.2
pillow==11.1.0
python-multipart==0.0.20
httpx==0.28.1
aiofiles==24.1.0
